+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This file is generated from git history by `php tools/release.php`.
 - Updated internal project files
 - Re-arraged the infobox in the dashboard for better user experience
 - Re-ordering infoBox for better user experience
+- Implementing progressive login lockout on the login page
 
 ### Fixed
 - Bug fix on change log and version number